refaktoroitu ItemQueues erillisiin komponentteihin.
oli haasteita tiettyjä osia koodista niin refaktoroitu nyt vähän selkeämmäksi
